Verify command (#1)
Adds a `verify` command that checks that the entire package-set successfully compiles. Using this one can build up a package-set by adding more and more packages and verifying that things still build in between. It's also a suitable command to be run on the CI of a package set maintained in a repository.