Add `vessel install --force` (#56)
Resolves #55 by adding an optional `--force` / `-f` flag to `vessel install`.

Since this feature requires behavior equivalent to `rm -rf`, this PR also fixes a potentially major vulnerability in the form of using unsanitized package names or versions containing `..`, `/`, or `~`, making it possible to modify (or delete) any directory on a user's machine. I fixed this by adding basic package name/version validation with the following rules:

- Package name/version must match the following regex: `[A-Za-z0-9.-_]+` (i.e. the [Portable Filename Character Set](https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/V1_chap03.html#tag_03_282))
- Package name/version cannot consist entirely of the `.` character
- Package name/version cannot start with a `-` character

This validation could break backwards-compatibility with some unusually-named packages, but this seems to be worth it for security reasons.